+++ src/VBox/Runtime/testcase/tstRTR0CommonDriver.h
@@ -4,7 +4,7 @@
  */
 
 /*
- * Copyright (C) 2010-2015 Oracle Corporation
+ * Copyright (C) 2010-2016 Oracle Corporation
  *
  * This file is part of VirtualBox Open Source Edition (OSE), as
  * available from http://www.virtualbox.org. This file is free software;
@@ -46,7 +46,7 @@ RTTEST              g_hTest;
 /** The test & service name. */
 char                g_szSrvName[64];
 /** The length of the service name.  */
-uint32_t            g_cchSrvName;
+size_t              g_cchSrvName;
 /** The base address of the service module. */
 void               *g_pvImageBase;
